Ryddet opp i pakkestruktur og navngiving.
PB-606. minesaker-api - Sette opp grapQL-klient mot SAF
