+++ src/main/kotlin/no/nav/personbruker/minesaker/api/saf/transformers/HentJournalposterTransformer.kt
@@ -1,26 +1,25 @@
-package no.nav.personbruker.minesaker.api.saf
+package no.nav.personbruker.minesaker.api.saf.transformers
 
-import no.nav.dokument.saf.selvbetjening.generated.dto.HentSakstemaer
+import no.nav.dokument.saf.selvbetjening.generated.dto.HentJournalposter
 import no.nav.personbruker.minesaker.api.common.exception.MissingFieldException
-import no.nav.personbruker.minesaker.api.common.exception.SafException
 import no.nav.personbruker.minesaker.api.saf.domain.MinimaltSakstema
 
-object HentSakstemaerTransformer {
+object HentJournalposterTransformer {
 
-    fun toInternal(externalTemaer: List<HentSakstemaer.Sakstema>): List<MinimaltSakstema> {
+    fun toInternal(externalTemaer: List<HentJournalposter.Sakstema>): List<MinimaltSakstema> {
         return externalTemaer.map { external ->
             toInternal(external)
         }
     }
 
-    fun toInternal(external: HentSakstemaer.Sakstema): MinimaltSakstema {
+    fun toInternal(external: HentJournalposter.Sakstema): MinimaltSakstema {
         return MinimaltSakstema(
             external.navn ?: throw MissingFieldException("navn"),
             external.kode ?: throw MissingFieldException("kode")
         )
     }
 
-    fun toInternal(externalData: HentSakstemaer.Result): List<MinimaltSakstema> {
+    fun toInternal(externalData: HentJournalposter.Result): List<MinimaltSakstema> {
         return externalData.dokumentoversiktSelvbetjening.tema.map { external ->
             toInternal(external)
         }
