+++ src/test/kotlin/no/nav/personbruker/minesaker/api/saf/requests/SakstemaerRequestTest.kt
@@ -1,6 +1,7 @@
 package no.nav.personbruker.minesaker.api.saf.requests
 
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
+import com.fasterxml.jackson.module.kotlin.jsonMapper
 import org.amshove.kluent.`should contain`
 import org.junit.jupiter.api.Test
 
@@ -12,18 +13,16 @@ internal class SakstemaerRequestTest {
 
     @Test
     fun `Skal bygge opp en korrekt sporring`() {
-        val expectedSakstema = "FOR"
         val expectedFields = "tema { navn kode }"
         val identAsQueryVarible = "\$ident : String!"
-        val expectedSakstemaAsInputVariable = """"temaetSomSkalHentes":"$expectedSakstema""""
+        val expectedAllSakstemar = """tema: []"""
 
-        val request = JournalposterRequest.create(dummyIdent, expectedSakstema)
-
+        val request = SakstemaerRequest.create(dummyIdent)
         val requestAsJson = objectMapper.writeValueAsString(request)
 
         requestAsJson `should contain` expectedFields
         requestAsJson `should contain` identAsQueryVarible
-        requestAsJson `should contain` expectedSakstemaAsInputVariable
+        requestAsJson `should contain` expectedAllSakstemar
     }
 
 }
