+++ src/test/kotlin/no/nav/personbruker/minesaker/api/saf/transformers/HentJournalposterTransformerTest.kt
@@ -1,4 +1,4 @@
-package no.nav.personbruker.minesaker.api.saf
+package no.nav.personbruker.minesaker.api.saf.transformers
 
 
 import no.nav.dokument.saf.selvbetjening.generated.dto.HentJournalposter
